Refactor code
All the download functions can be represented by a single function that downloads the CSVs as requested by the user. How to parameterize the data request can be alterered ultimately, but this approach should make the package as a whole more easy to maintain and extend.

A few internal functions are also created for the parsing and building of the URLs and for the retrieval of some data internally.

Speaking of data, the various request parameters are now tucked away into a data structure (currently a named vector). The script for generating the data is included in the package in the data folder.

Finally, the documentation has been updated, thought still in draft form. This includes DESCRIPTION, NAMESPACE, and help file(s).