Upgrade the documentation
The data source for the search term values was created and stored in as internal data, with the source code stored in the 'data-raw' directory (which is added to .Rbuildignore).

Also in writing the documentation, and to ease maintenance, an Rmd file is created that will be sourced from the roxygen comments of the main function. The aim is to provide a guide to users on how to call the function to access the data.